Moved trigger and blacklister to ssadmin
This commit moves the in-container admin functions to a new admin
executable. In particular, trigger and blacklisting functionality
has been moved into a single ssadmin command in the controller
image. All necessary changes to dockerfile, Makefile and gitignore
are included.
