+++ .gitignore
@@ -14,11 +14,13 @@
 .glide/
 
 /controller
+/ssadmin
 /kubeseal
 /controller.image
 /*-static
 /controller.yaml
 /sealedsecret-crd.yaml
 /docker/controller
+/docker/ssadmin
 *.iml
 .idea
